+++ cpp/turbodbc/Library/src/cursor.cpp
@@ -1,6 +1,8 @@
 #include <turbodbc/cursor.h>
 #include <turbodbc/make_description.h>
 
+#include <turbodbc/buffer_size.h>
+
 #include <cpp_odbc/statement.h>
 #include <cpp_odbc/error.h>
 
@@ -15,11 +17,11 @@
 namespace turbodbc {
 
 cursor::cursor(std::shared_ptr<cpp_odbc::connection const> connection,
-               std::size_t rows_to_buffer,
+               turbodbc::buffer_size buffer_size,
                std::size_t parameter_sets_to_buffer,
                bool use_async_io) :
 	connection_(connection),
-	rows_to_buffer_(rows_to_buffer),
+	buffer_size_(buffer_size),
 	parameter_sets_to_buffer_(parameter_sets_to_buffer),
 	use_async_io_(use_async_io),
 	query_()
@@ -34,7 +36,7 @@ void cursor::prepare(std::string const & sql)
 	query_.reset();
 	auto statement = connection_->make_statement();
 	statement->prepare(sql);
-	query_ = std::make_shared<query>(statement, rows_to_buffer_, parameter_sets_to_buffer_, use_async_io_);
+	query_ = std::make_shared<query>(statement, buffer_size_, parameter_sets_to_buffer_, use_async_io_);
 }
 
 void cursor::execute()