+++ cpp/turbodbc/Library/turbodbc/query.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <turbodbc/buffer_size.h>
+
 #include <cpp_odbc/statement.h>
 #include <turbodbc/result_sets/result_set.h>
 #include <turbodbc/parameter_sets/field_parameter_set.h>
@@ -12,7 +14,7 @@ namespace turbodbc {
 class query {
 public:
 	query(std::shared_ptr<cpp_odbc::statement const> statement,
-		  std::size_t rows_to_buffer,
+		  turbodbc::buffer_size buffer_size,
 		  std::size_t parameter_sets_to_buffer,
 		  bool use_double_buffering);
 
@@ -27,7 +29,7 @@ public:
 private:
 	std::shared_ptr<cpp_odbc::statement const> statement_;
 	field_parameter_set parameters_;
-	std::size_t rows_to_buffer_;
+	turbodbc::buffer_size buffer_size_;
 	bool use_double_buffering_;
 	std::shared_ptr<result_sets::result_set> results_;
 };