+++ python/turbodbc/connect.py
@@ -1,6 +1,7 @@
 from __future__ import absolute_import
 
 from turbodbc_intern import connect as intern_connect
+from turbodbc_intern import Rows
 
 from .exceptions import translate_exceptions
 from .connection import Connection
@@ -12,7 +13,7 @@ def _make_connection_string(dsn, **kwargs):
 
 
 @translate_exceptions
-def connect(dsn=None, rows_to_buffer=None, parameter_sets_to_buffer=None, use_async_io=False, **kwargs):
+def connect(dsn=None, read_buffer_size=None, rows_to_buffer=None, parameter_sets_to_buffer=None, use_async_io=False, **kwargs):
     """
     Create a connection with the database identified by the dsn
     :param dsn: data source name as given in the odbc.ini file
@@ -36,7 +37,10 @@ def connect(dsn=None, rows_to_buffer=None, parameter_sets_to_buffer=None, use_as
     connection = Connection(intern_connect(_make_connection_string(dsn, **kwargs)))
 
     if rows_to_buffer:
-        connection.impl.rows_to_buffer = rows_to_buffer
+        connection.impl.set_buffer_size(Rows(rows_to_buffer))
+
+    if read_buffer_size:
+        connection.impl.set_buffer_size(read_buffer_size)
 
     if parameter_sets_to_buffer:
         connection.impl.parameter_sets_to_buffer = parameter_sets_to_buffer